refactor: move vllm server logic to a new module and update related functions
