+++ src/automation/tasks/scripts/guidellm_script.py
@@ -1,83 +1,10 @@
-import subprocess
-import requests
-import time
+
 import os
-import sys
-from urllib.parse import urlparse
 from clearml import Task
-import torch
-from automation.utils import resolve_model_id, cast_args
-import psutil
+from automation.utils import resolve_model_id, cast_args, kill_process_tree
+from automation.vllm import start_vllm_server
 from pyhocon import ConfigFactory
 
-
-SERVER_LOG_PREFIX = "vllm_server_log"
-
-
-def kill_process_tree(pid):
-    try:
-        parent = psutil.Process(pid)
-        children = parent.children(recursive=True)
-        for child in children:
-            child.terminate()  # or child.kill()
-        parent.terminate()
-    except psutil.NoSuchProcess:
-        pass
-
-
-def start_vllm_server(
-    vllm_args, 
-    model_id, 
-    target, 
-    server_wait_time, 
-    suffix,
-):
-    executable_path = os.path.dirname(sys.executable)
-    vllm_path = os.path.join(executable_path, "vllm")
-
-    num_gpus = torch.cuda.device_count()
-
-    parsed_target = urlparse(target)
-
-    server_command = [
-        f"{vllm_path}", "serve", 
-        model_id,
-        "--host", parsed_target.hostname, 
-        "--port", str(parsed_target.port),
-        "--tensor-parallel-size", str(num_gpus)
-    ]
-
-    subprocess_env = os.environ.copy()
-
-    for k, v in vllm_args.items():
-        if k.startswith("VLLM_"):
-            subprocess_env[k] = str(v)
-        else:
-            if v == True or v == "True":
-                v = "true"
-            server_command.extend([f"--{k}", str(v)])
-
-    server_log_file = open(f"{SERVER_LOG_PREFIX}_{suffix}.txt", "w")
-    server_process = subprocess.Popen(server_command, stdout=server_log_file, stderr=server_log_file, shell=False, env=subprocess_env)
-
-    delay = 5
-    server_initialized = False
-    for _ in range(server_wait_time // delay):
-        try:
-            response = requests.get(target + "/models")
-            if response.status_code == 200:
-                print("Server initialized")
-                server_initialized = True
-                break  # Exit the loop if the request is successful
-        except requests.exceptions.RequestException as e:
-            pass
-
-        time.sleep(delay)
-
-    if server_initialized:
-        return server_process, True
-    else:
-        return server_process, False
 
 def main():
     task = Task.current_task()
@@ -106,7 +33,7 @@ def main():
     model_id = resolve_model_id(args["Args"]["model_id"], bool(args["Args"]["clearml_model"]), task)
 
     # Start vLLM server
-    server_process, server_initialized = start_vllm_server(
+    server_process, server_initialized, server_log = start_vllm_server(
         vllm_args,
         model_id,
         guidellm_args["target"],
@@ -116,7 +43,7 @@ def main():
 
     if not server_initialized:
         kill_process_tree(server_process.pid)
-        task.upload_artifact(name="vLLM server log", artifact_object=f"{SERVER_LOG_PREFIX}_{task.id}.txt")
+        task.upload_artifact(name="vLLM server log", artifact_object=server_log)
         raise AssertionError("Server failed to intialize")
 
     # Parse through environment variables
@@ -131,7 +58,7 @@ def main():
     kill_process_tree(server_process.pid)
 
     task.upload_artifact(name="guidellm guidance report", artifact_object=report.to_json())
-    task.upload_artifact(name="vLLM server log", artifact_object=f"{SERVER_LOG_PREFIX}_{task.id}.txt")
+    task.upload_artifact(name="vLLM server log", artifact_object=server_log)
 
 if __name__ == '__main__':
     main()
