+++ src/automation/vllm/__init__.py
@@ -1,0 +1,1 @@
+from automation.vllm.server import start_vllm_server
